requirements.txt
streamlit==1.24.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
gradio

------------------------------------------------------------------------------------
			RUN THIS CODE IN CMD AS ADMIN 👇👇
		
			pip install -r requirements.txt
-------------------------------------------------------------------------------------
